+++ src/index.jsx
@@ -1,25 +1,4 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
-import ToDo from './ToDo';
-
-const App = () => {
-  const [list, setList] = React.useState([]);
-
-  const onClickAppendToDo = (text) => {
-    setList([...list, { id: Date.now() + text, text }]);
-  };
-
-  const onClickRemoveToDo = (id) => {
-    setList(list.filter((item) => item.id !== id));
-  };
-
-  return (
-    <ToDo
-      appendHandler={onClickAppendToDo}
-      removeHandler={onClickRemoveToDo}
-      list={list}
-    />
-  );
-};
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
